+++ compiler/frontend/ast_uncurry_gen.ml
@@ -25,7 +25,7 @@
 open Ast_helper
 
 (* Handling `fun [@this]` used in `object [@bs] end` *)
-let to_method_callback loc (self : Bs_ast_mapper.mapper) label
+let to_method_callback ~async loc (self : Bs_ast_mapper.mapper) label
     (self_pat : Parsetree.pattern) body : Parsetree.expression_desc =
   let self_pat = self.pat self self_pat in
   (match Ast_pat.is_single_variable_pattern_conservative self_pat with
@@ -50,7 +50,9 @@ let to_method_callback loc (self : Bs_ast_mapper.mapper) label
   let arity = List.length rev_extra_args in
   let body =
     match body.pexp_desc with
-    | Pexp_fun f -> {body with pexp_desc = Pexp_fun {f with arity = Some arity}}
+    | Pexp_fun f ->
+      Ast_async.make_function_async ~async
+        {body with pexp_desc = Pexp_fun {f with arity = Some arity; async}}
     | _ -> body
   in
   let arity_s = string_of_int arity in