+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 - Fix rewatch not recompiling on changes under windows. https://github.com/rescript-lang/rescript/pull/7690
 - Fix locations of regex literals. https://github.com/rescript-lang/rescript/pull/7683
 - Fix async React component compilation. https://github.com/rescript-lang/rescript/pull/7704
+- Fix @this with async keyword. https://github.com/rescript-lang/rescript/pull/7702
 
 # 12.0.0-beta.2